More module rationalization
Shift everything that's entirely generic widgets into the widgets
sub-module, and everything application specific into the top-level
monitor module.
